To add as a new unit:
Go to the uncompiled_lua .scd file in the game's local files. Then go to Units -> Illuminate -> UIX0111. Copy all files. In your mod.scd folder, add a folder named "Units", and inside that folder, a folder named "Illuminate", inside the 'Illuminate' folder, add another folder, with any name, for this tutorial, I will name the folder "UIX0111A", I would suggest conforming to the naming system, however. Add the files from the UIX0111 unit in your equivalent of the "UIX0111A" folder. Rename all files, by replacing the "UIX0111" with the name of your folder, in this case "UIX0111A", so for example the .bp file should be named (in his case) "UIX0111A_unit.bp". Open the blueprint file, and manually apply the above changes. Most of it is just number tweaking. If done successfully, you should have a new unit in the Illuminate Experimental gantry, its icon will be black - this is because one was never created. More on this in a future episode.